+++ keep_screen_on/example/lib/main.dart
@@ -16,8 +16,11 @@ class MyApp extends StatefulWidget {
 class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
 
   bool? _isKeepScreenOn;
+  bool? _isAllowLockWhileScreenOn;
   DateTime? _changeAt;
   DateTime? _pausedAt;
+
+  final _appLifecycleStateHistories = <AppLifecycleStateHistory>[];
 
   @override
   void initState() {
@@ -35,6 +38,18 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
 
   @override
   void didChangeAppLifecycleState(AppLifecycleState state) {
+    final stateHistory = AppLifecycleStateHistory(
+      time: DateTime.now(),
+      state: state,
+    );
+    if (context.mounted) {
+      setState(() {
+        _appLifecycleStateHistories.add(stateHistory);
+      });
+    } else {
+      _appLifecycleStateHistories.add(stateHistory);
+    }
+    
     if (state == AppLifecycleState.paused) {
       setState(() {
         _pausedAt = DateTime.now();
@@ -44,6 +59,7 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
 
   @override
   Widget build(BuildContext context) {
+
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
@@ -52,6 +68,7 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
         body: Padding(
           padding: const EdgeInsets.all(16.0),
           child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
             children: [
               ListView(
                 shrinkWrap: true,
@@ -59,6 +76,10 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
                   ListTile(
                     title: Text('Keep screen on'),
                     subtitle: (_isKeepScreenOn == null) ? Text('Unknown') : Text(_isKeepScreenOn.toString()),
+                  ),
+                  ListTile(
+                    title: Text('FLAG_ALLOW_LOCK_WHILE_SCREEN_ON'),
+                    subtitle: (_isAllowLockWhileScreenOn == null) ? Text('Unknown') : Text(_isAllowLockWhileScreenOn.toString()),
                   ),
                   ListTile(
                     title: Text('Change date and time'),
@@ -75,43 +96,86 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
                 ],
               ),
 
-              ButtonBar(
-                alignment: MainAxisAlignment.center,
+              Wrap(
+                spacing: 8.0,
+                runSpacing: 8.0,
                 children: [
                   ElevatedButton.icon(
-                    icon: Icon(Icons.arrow_upward),
-                    label: Text('Turn on'),
+                    icon: Icon(Icons.lightbulb),
+                    label: Text('Turn on and allowLockWhileScreenOn'),
                     onPressed: () {
-                      KeepScreenOn.turnOn(true).then((value) async {
-                        final isOn = await KeepScreenOn.isOn;
-                        setState(() {
-                          _isKeepScreenOn = isOn;
-                          _changeAt = DateTime.now();
-                        });
-                      });
+                      KeepScreenOn.turnOn(withAllowLockWhileScreenOn: true)
+                          .then((value) => _refreshState());
                     },
                   ),
 
                   ElevatedButton.icon(
-                    icon: Icon(Icons.arrow_downward),
+                    icon: Icon(Icons.lightbulb_outline),
+                    label: Text('Turn off and allowLockWhileScreenOn'),
+                    onPressed: () {
+                      KeepScreenOn.turnOff(withAllowLockWhileScreenOn: true)
+                          .then((value) => _refreshState());
+                    },
+                  ),
+
+                  ElevatedButton.icon(
+                    icon: Icon(Icons.lightbulb),
+                    label: Text('Turn on'),
+                    onPressed: () {
+                      KeepScreenOn.turnOn().then((value) => _refreshState());
+                    },
+                  ),
+
+                  ElevatedButton.icon(
+                    icon: Icon(Icons.lightbulb_outline),
                     label: Text('Turn off'),
                     onPressed: () {
-                      KeepScreenOn.turnOn(false).then((value) async {
-                        final isOn = await KeepScreenOn.isOn;
-                        setState(() {
-                          _isKeepScreenOn = isOn;
-                          _changeAt = DateTime.now();
-                        });
-                      });
+                      KeepScreenOn.turnOff().then((value) => _refreshState());
                     },
                   ),
                 ],
+              ),
+
+              const Text('AppLifecycleState'),
+              Flexible(
+                child: ListView.builder(
+                  reverse: true,
+                  itemBuilder: (context, index) {
+                    final item = _appLifecycleStateHistories.elementAt(index);
+
+                    final leadingStyle = DefaultTextStyle.of(context).style.copyWith(
+                      color: item.stateColor,
+                      fontWeight: FontWeight.bold,
+                    );
+
+                    return ListTile(
+                      leading: Text(item.state.name, style: leadingStyle),
+                      title: Text(item.time.toLocal().toString(), textAlign: TextAlign.end),
+                      dense: true,
+                    );
+                  },
+                  itemCount: _appLifecycleStateHistories.length,
+                ),
               ),
             ],
           ),
         ),
       ),
     );
+  }
+
+  Future<void> _refreshState() {
+
+    return Future.wait([
+      KeepScreenOn.isOn,
+      KeepScreenOn.isAllowLockWhileScreenOn,
+    ]).then((values) {
+      setState(() {
+        _isKeepScreenOn = values.elementAtOrNull(0) ?? false;
+        _isAllowLockWhileScreenOn = values.elementAtOrNull(1) ?? false;
+        _changeAt = DateTime.now();
+        });
+    });
   }
 }
 
@@ -164,5 +228,32 @@ class _ElapsedTimeTextState extends State<ElapsedTimeText> {
 
   static String formatTwoDigits(int num) {
     return num.toString().padLeft(2, '0');
+  }
+}
+
+class AppLifecycleStateHistory {
+  final DateTime time;
+  final AppLifecycleState state;
+
+  AppLifecycleStateHistory({
+    required this.time,
+    required this.state,
+  });
+
+  Color get stateColor {
+    switch (state) {
+      case AppLifecycleState.detached:
+        return Colors.red;
+      case AppLifecycleState.resumed:
+        return Colors.green;
+      case AppLifecycleState.inactive:
+        return Colors.grey;
+      case AppLifecycleState.hidden:
+        return Colors.orange;
+      case AppLifecycleState.paused:
+        return Colors.indigo;
+      default:
+        return Colors.black;
+    }
   }
 }